added functions into rpchelper
