+++ turbo/rpchelper/rpc_block.go
@@ -1,4 +1,4 @@
-package commands
+package rpchelper
 
 import (
 	"fmt"
@@ -7,48 +7,9 @@ import (
 	"github.com/ledgerwatch/erigon/common"
 	"github.com/ledgerwatch/erigon/core/rawdb"
 	"github.com/ledgerwatch/erigon/eth/stagedsync/stages"
-	"github.com/ledgerwatch/erigon/rpc"
 )
 
-func getBlockNumber(number rpc.BlockNumber, tx kv.Tx) (uint64, error) {
-	var blockNum uint64
-	var err error
-	switch number {
-
-	case rpc.LatestBlockNumber:
-		blockNum, err = getLatestBlockNumber(tx)
-		if err != nil {
-			return 0, err
-		}
-	case rpc.PendingBlockNumber:
-		blockNum, err = getLatestBlockNumber(tx)
-		if err != nil {
-			return 0, err
-		}
-
-	case rpc.EarliestBlockNumber:
-		blockNum = 0
-
-	case rpc.FinalizeBlockNumber:
-		blockNum, err = getFinalizedBlockNumber(tx)
-		if err != nil {
-			return 0, err
-		}
-
-	case rpc.SafeBlockNumber:
-		blockNum, err = getSafeBlockNumber(tx)
-		if err != nil {
-			return 0, err
-		}
-
-	default:
-		blockNum = uint64(number.Int64())
-	}
-
-	return blockNum, nil
-}
-
-func getLatestBlockNumber(tx kv.Tx) (uint64, error) {
+func GetLatestBlockNumber(tx kv.Tx) (uint64, error) {
 	forkchoiceHeadHash := rawdb.ReadForkchoiceHead(tx)
 	if forkchoiceHeadHash != (common.Hash{}) {
 		forkchoiceHeadNum := rawdb.ReadHeaderNumber(tx, forkchoiceHeadHash)
@@ -65,7 +26,7 @@ func getLatestBlockNumber(tx kv.Tx) (uint64, error) {
 	return blockNum, nil
 }
 
-func getFinalizedBlockNumber(tx kv.Tx) (uint64, error) {
+func GetFinalizedBlockNumber(tx kv.Tx) (uint64, error) {
 	forkchoiceFinalizedHash := rawdb.ReadForkchoiceFinalized(tx)
 	if forkchoiceFinalizedHash != (common.Hash{}) {
 		forkchoiceFinalizedNum := rawdb.ReadHeaderNumber(tx, forkchoiceFinalizedHash)
@@ -77,7 +38,7 @@ func getFinalizedBlockNumber(tx kv.Tx) (uint64, error) {
 	return 0, nil
 }
 
-func getSafeBlockNumber(tx kv.Tx) (uint64, error) {
+func GetSafeBlockNumber(tx kv.Tx) (uint64, error) {
 	forkchoiceSafeHash := rawdb.ReadForkchoiceSafe(tx)
 	if forkchoiceSafeHash != (common.Hash{}) {
 		forkchoiceSafeNum := rawdb.ReadHeaderNumber(tx, forkchoiceSafeHash)
